feat: added a register_actions helper method and moved get_files to the new folder
